# openapi.yaml
openapi: 3.0.0
info:
  title: Drone Backend API
  description: Smart Surveillance Platform - Drone Video Backend Service
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/recordings:
    post:
      tags:
        - Recordings
      summary: Create recording entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cameraId
                - startAt
                - endAt
                - sizeBytes
              properties:
                cameraId:
                  type: string
                  example: "cam-001"
                startAt:
                  type: string
                  format: date-time
                  example: "2026-06-05T10:00:00Z"
                endAt:
                  type: string
                  format: date-time
                  example: "2026-06-05T11:00:00Z"
                sizeBytes:
                  type: integer
                  example: 5000000
                storageUrl:
                  type: string
                  example: "s3://bucket/recordings/rec-001"
      responses:
        '201':
          description: Recording created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'

    get:
      tags:
        - Recordings
      summary: List recordings
      security:
        - bearerAuth: []
      parameters:
        - name: cameraId
          in: query
          schema:
            type: string
          description: Filter by camera ID
        - name: start
          in: query
          schema:
            type: string
            format: date-time
          description: Start date for range query
        - name: end
          in: query
          schema:
            type: string
            format: date-time
          description: End date for range query
      responses:
        '200':
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recording'
  /api/recordings/{id}:
    get:
      tags:
        - Recordings
      summary: Get recording by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '404':
          description: Recording not found

    put:
      tags:
        - Recordings
      summary: Update recording
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recording'
      responses:
        '200':
          description: Recording updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '404':
          description: Recording not found

    delete:
      tags:
        - Recordings
      summary: Delete recording
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Recording deleted
        '404':
          description: Recording not found
  /api/recordings/{id}/metadata:
    get:
      tags:
        - Recordings
      summary: Get recording metadata
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
  /api/recordings/{id}/generate-playback-url:
    post:
      tags:
        - Recordings
      summary: Generate playback URL
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                protocol:
                  type: string
                  enum: [hls, dash]
                  example: "hls"
                ttl:
                  type: integer
                  description: Time to live in seconds
                  example: 300
      responses:
        '200':
          description: Playback URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  playbackUrl:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /api/recordings/timelines/{siteId}:
    get:
      tags:
        - Recordings
      summary: Get recording timeline for site
      security:
        - bearerAuth: []
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Timeline data
          content:
            application/json:
              schema:
                type: object
                properties:
                  siteId:
                    type: string
                  range:
                    type: object
                    properties:
                      from:
                        type: string
                        format: date-time
                      to:
                        type: string
                        format: date-time
                  tracks:
                    type: array
                    items:
                      type: object
                      properties:
                        cameraId:
                          type: string
                        segments:
                          type: array
                          items:
                            type: object
                            properties:
                              start:
                                type: string
                                format: date-time
                              end:
                                type: string
                                format: date-time
                              thumbnail:
                                type: string
                              recordingId:
                                type: string
...
  /api/streams/available:
    get:
      tags:
        - Streams
      summary: List available streams
      security:
        - bearerAuth: []
      parameters:
        - name: cameraId
          in: query
          required: true
          schema:
            type: string
        - name: profile
          in: query
          schema:
            type: string
            example: "720p"
      responses:
        '200':
          description: Available streams
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    protocol:
                      type: string
                    profile:
                      type: string
                    latencyMs:
                      type: integer
  /api/streams/watch:
    post:
      tags:
        - Streams
      summary: Start watching live stream
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cameraId
                - protocol
              properties:
                cameraId:
                  type: string
                profile:
                  type: string
                  example: "720p"
                protocol:
                  type: string
                  enum: [webrtc, hls]
                preferLowLatency:
                  type: boolean
      responses:
        '200':
          description: Stream session started
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      sessionId:
                        type: string
                      webrtcOffer:
                        type: object
                        properties:
                          sdp:
                            type: string
                      signalingUrl:
                        type: string
                  - type: object
                    properties:
                      playbackUrl:
                        type: string
  /api/streams/stop:
    post:
      tags:
        - Streams
      summary: Stop watching stream
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
      responses:
        '200':
          description: Stream stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  stopped:
                    type: boolean
  /api/streams/subscribe:
    post:
      tags:
        - Streams
      summary: Subscribe to multiple camera streams
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cameraIds
              properties:
                sessionId:
                  type: string
                cameraIds:
                  type: array
                  items:
                    type: string
                profiles:
                  type: array
                  items:
                    type: string
                layout:
                  type: string
                  enum: [grid, custom]
      responses:
        '200':
          description: Subscription created
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptionId:
                    type: string
                  streams:
                    type: array
                    items:
                      type: object
                      properties:
                        cameraId:
                          type: string
                        playbackUrl:
                          type: string
...
components:
  schemas:
...
    Recording:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cameraId:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        sizeBytes:
          type: integer
        storageUrl:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT